path_planner_3d/requirements.txt
numpy>=1.24.0,<2.5.0
PyQt5==5.15.11
PyQt5_sip==12.17.0
pyqtgraph>=0.13.0
PyOpenGL>=3.1.0
